requirements.txt
#
####### requirements.txt #######
#
###### Requirements without Version Specifiers ######
Pillow
#
